requirements.txt
backoff>=2.2.1
cut-cross-entropy
datasets==2.21.0
dotenv>=0.9.9
hatchling>=1.27.0
huggingface-hub[hf-transfer]>=0.28.1
ipykernel>=6.29.5
matplotlib>=3.10.0
nbformat>=5.10.4
nnsight
openai>=1.63.0
peft>=0.14.0
scikit-learn>=1.6.1
simple-parsing>=0.1.6
torchtyping>=0.1.5
transformers>=4.50.0
vllm==0.8.2
wandb>=0.19.5
pydantic>=2.11.4
git+https://github.com/davidbau/baukit
git+https://github.com/cadentj/autointerp.git@dash
